+++ frontend/src/views/Community/FAQ.vue
@@ -9,7 +9,7 @@
     </v-data-table>
     <v-row justify="end" class="mt-2">
         <v-col cols="auto">
-            <v-btn :to="'/community/writePost/faq'">글쓰기</v-btn>
+            <v-btn :to="'/writePost/faq'">글쓰기</v-btn>
         </v-col>
     </v-row>
 </v-container>
