+++ frontend/src/views/Community/Notice.vue
@@ -19,7 +19,7 @@
             </v-row>
         </v-col>
         <v-col cols="auto">
-            <v-btn :to="'/community/writePost/notice'">글쓰기</v-btn>
+            <v-btn :to="'/writePost/notice'">글쓰기</v-btn>
         </v-col>
     </v-row>
 </v-container>
